pyproject.toml
[project]
name = "doc-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2pdf>=0.1.8",
    "fastapi>=0.120.2",
    "gunicorn>=23.0.0",
    "mammoth>=1.11.0",
    "openai>=2.6.1",
    "pip>=25.3",
    "pymupdf>=1.26.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-magic-bin==0.4.14",
    "python-multipart>=0.0.20",
    "spacy>=3.8.7",
    "uvicorn>=0.38.0",
    "xhtml2pdf>=0.2.17",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "httpx",
]